Refactoring the rqt_rotors HIL plugin
