+++ rqt_rotors/package.xml
@@ -17,20 +17,14 @@
   <!-- Dependencies needed to compile this package. -->
   <build_depend>mavros_msgs</build_depend>
   <build_depend>rospy</build_depend>
-  <build_depend>rotors_comm</build_depend>
   <build_depend>rqt_gui</build_depend>
   <build_depend>rqt_gui_py</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>std_srvs</build_depend>
 
   <!-- Dependencies needed after this package is compiled. -->
   <run_depend>mavros_msgs</run_depend>
   <run_depend>rospy</run_depend>
-  <run_depend>rotors_comm</run_depend>
   <run_depend>rqt_gui</run_depend>
   <run_depend>rqt_gui_py</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>std_srvs</run_depend>
 
   <export>
     <rqt_gui plugin="${prefix}/plugin.xml"/>
