+++ rqt_rotors/plugin.xml
@@ -1,16 +1,12 @@
 <library path="src">
-  <class name="Rotors" type="rqt_rotors.rotors.Rotors" base_class_type="rqt_gui_py::Plugin">
+  <class name="Hil Plugin" type="rqt_rotors.hil_plugin.HilPlugin" base_class_type="rqt_gui_py::Plugin">
     <description>
-      Plugin for RotorS.
+      Plugin for RotorS HIL interface.
     </description>
     <qtgui>
-      <group>
-        <label>ASL</label>        
-        <icon type="theme">folder</icon>
-      </group>
-      <label>RotorS</label>
+      <label>RotorS HIL</label>
       <icon type="theme">applications-other</icon>
-      <statustip>Plugin for RotorS.</statustip>
+      <statustip>Plugin for RotorS HIL interface.</statustip>
     </qtgui>
   </class>
 </library>
