+++ rqt_rotors/resource/HilPlugin.ui
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>HilWidget</class>
- <widget class="QWidget" name="HilWidget">
+ <class>rotors_hil_gui</class>
+ <widget class="QWidget" name="rotors_hil_gui">
   <property name="geometry">
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>400</width>
+    <width>300</width>
     <height>500</height>
    </rect>
   </property>
@@ -18,8 +18,8 @@
   </property>
   <property name="minimumSize">
    <size>
-    <width>400</width>
-    <height>500</height>
+    <width>300</width>
+    <height>450</height>
    </size>
   </property>
   <property name="windowTitle">
@@ -41,7 +41,7 @@ QGroupBox::title {
    </property>
    <property name="geometry">
     <rect>
-     <x>20</x>
+     <x>30</x>
      <y>20</y>
      <width>241</width>
      <height>301</height>
@@ -270,7 +270,7 @@ QGroupBox::title {
   <widget class="QWidget" name="verticalLayoutWidget">
    <property name="geometry">
     <rect>
-     <x>20</x>
+     <x>70</x>
      <y>340</y>
      <width>160</width>
      <height>95</height>
@@ -308,71 +308,6 @@ QGroupBox::title {
      </widget>
     </item>
    </layout>
-  </widget>
-  <widget class="QWidget" name="verticalLayoutWidget_2">
-   <property name="geometry">
-    <rect>
-     <x>210</x>
-     <y>340</y>
-     <width>161</width>
-     <height>131</height>
-    </rect>
-   </property>
-   <layout class="QVBoxLayout" name="verticalLayout_2">
-    <item>
-     <widget class="QPushButton" name="button_reconstruct">
-      <property name="enabled">
-       <bool>true</bool>
-      </property>
-      <property name="text">
-       <string>Start Reconstruction</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="button_reset_model">
-      <property name="enabled">
-       <bool>true</bool>
-      </property>
-      <property name="text">
-       <string>Reset Model</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="button_set_position_rate">
-      <property name="enabled">
-       <bool>true</bool>
-      </property>
-      <property name="text">
-       <string>Set Position Rate</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="button_set_imu_rate">
-      <property name="text">
-       <string>Set IMU Rate</string>
-      </property>
-     </widget>
-    </item>
-   </layout>
-  </widget>
-  <widget class="QPushButton" name="button_record_rosbag">
-   <property name="enabled">
-    <bool>true</bool>
-   </property>
-   <property name="geometry">
-    <rect>
-     <x>20</x>
-     <y>440</y>
-     <width>158</width>
-     <height>27</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>Record Bag</string>
-   </property>
   </widget>
  </widget>
  <resources/>
