+++ rqt_rotors/setup.py
@@ -6,6 +6,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 setup_args = generate_distutils_setup(
     packages=['rqt_rotors'],
     package_dir={'': 'src'},
+    scripts=['scripts/hil_plugin']
 )
 
 setup(**setup_args)
