--- rqt_rotors/src/rqt_rotors/rotors.py
@@ -1,53 +1,0 @@
-#!/usr/bin/env python
-import os
-import rospy
-from qt_gui.plugin import Plugin
-from argparse import ArgumentParser
-
-from .hil_widget import HilWidget
-
-class Rotors(Plugin):
-
-    def __init__(self, context):
-        super(Rotors, self).__init__(context)
-        # Give QObjects reasonable names
-        self.setObjectName('RotorsWidget')
-
-        # Process standalone plugin command-line arguments
-        parser = ArgumentParser()
-        # Add argument(s) to the parser.
-        parser.add_argument("-q", "--quiet", action="store_true",
-                      dest="quiet",
-                      help="Put plugin in silent mode")
-        args, unknowns = parser.parse_known_args(context.argv())
-        if not args.quiet:
-            print 'arguments: ', args
-            print 'unknowns: ', unknowns
-
-        # Create QWidget
-        self._widget = HilWidget();
-
-        # Show _widget.windowTitle on left-top of each plugin (when 
-        # it's set in _widget). This is useful when you open multiple 
-        # plugins at once. Also if you open multiple instances of your 
-        # plugin at once, these lines add number to make it easy to 
-        # tell from pane to pane.
-        if context.serial_number() > 1:
-            self._widget.setWindowTitle(self._widget.windowTitle() + (' (%d)' % context.serial_number()))
-
-        # Add widget to the user interface
-        context.add_widget(self._widget)
-
-    def shutdown_plugin(self):
-        # TODO unregister all publishers here
-        pass
-
-    def save_settings(self, plugin_settings, instance_settings):
-        # TODO save intrinsic configuration, usually using:
-        # instance_settings.set_value(k, v)
-        pass
-
-    def restore_settings(self, plugin_settings, instance_settings):
-        # TODO restore intrinsic configuration, usually using:
-        # v = instance_settings.value(k)
-        pass
